Uses unsigned long long now

+++ clock.c
@@ -10,7 +10,7 @@
 #define CLOCK_FREQ 1193180 // this is the same for all Intel CPUs
 #include "gamma.h"
 #include <stdint.h>
-uint64_t tick; // hundredths of a second
+unsigned long long tick; // hundredths of a second
 void clock_tick(registers_t regs)
 {
   ++tick;
